change MULTILIBS to use ?= so they can be overridden
the multilib setting have already been moved to the machine
configure file. according to the including order in bitbake.conf
the machine configure file is after local.conf file.so the value
of MULTILIBS will be overridden when we use external toolchain.
change MULTILIBS to use ?= to ensure external toolchain can work.

Signed-off-by: Zongchun Yu <Zongchun.Yu@freescale.com>
Change-Id: Ic40dd2964cc2048b1546bd8a18ed49e178b5dbe8
Reviewed-on: http://git.am.freescale.net:8181/24462
Reviewed-by: Ting Liu <ting.liu@freescale.com>
Tested-by: Ting Liu <ting.liu@freescale.com>