multilib: use ?= to allow overridden in local.conf
The multilib setting was set in <core>.inc which was included by
the <machine>.conf. In bitbake.conf, <machine>.conf is included
after local.conf. Change to use ?= to not override the value in
local.conf.

This is useful when using external toolchain which does not need
to enable multilib.

Signed-off-by: Zongchun Yu <Zongchun.Yu@freescale.com>