openssl: remove native dependency on cryptodev
The original intent of the code was good but the comment was missleading
since the cryptodev dependency was always set, even for the native package.

Dependency on acceleration engines (including cryptodev) should be disabled
for the native package to reduce the risk of incorrect builds and other
runtime issues. In particular, the chain of dependencies for such a high
profile crypto package should be minimal to help with code audit.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: I6bb03fb21e1d4b50cd8d85538e934e04be972781
Reviewed-on: http://git.am.freescale.net:8181/35578
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>